[400. Nth Digit][Accepted]committed by LeetCode Extension

+++ 400-Nth-Digit/solution.java
@@ -2,15 +2,15 @@ public class Solution {
     public int findNthDigit(int n) {
         if (n < 10) return n;
         int step = 1;
-        int base = 1;
-        while ((long)step * base * 9  < (long)n) {
+        long base = 1;
+        while (step * base * 9  < n) {
             n -= step * base * 9;
             step++;
             base *=10;
         }
         n--;
         base += n / step;
-        String s = Integer.toString(base);
+        String s = Long.toString(base);
 		return Character.getNumericValue(s.charAt((n) % step));
     }
 }
